update and add files

+++ index.php
@@ -27,15 +27,17 @@ include "components/header.php";
                         <?= $menu['title'] ?>
                     </p>
                 </div>
-                <div class="card-footer d-flex flex-wrap justify-content-between align-items-center">
-                    <p> $ <?= $menu['price'] ?></p>
-                    <p><i class="bi bi-cart-plus-fill"></i></p>
-                    <p class="w-100">
-                        <a href="menu-show.php?id=<?= $menu['id'] ?>" class="btn btn-link border me-2"><i class="bi bi-eye-fill"></i></a>
-                        <a href="menu-edit.php?id=<?= $menu['id'] ?>" class="btn btn-link border me-2"><i class="bi bi-pencil-square"></i></a>
-                        <a href="actions/menu-delete.php?id=<?= $menu['id'] ?>" class="btn btn-link border me-2"><i class="bi bi-trash"></i></a>
-                    </p>
-                </div>
+                <?php if (isset($_SESSION['admin'])) : ?>
+                    <div class="card-footer d-flex flex-wrap justify-content-between align-items-center">
+                        <p> $ <?= $menu['price'] ?></p>
+                        <p><i class="bi bi-cart-plus-fill"></i></p>
+                        <p class="w-100">
+                            <a href="menu-show.php?id=<?= $menu['id'] ?>" class="btn btn-link border me-2"><i class="bi bi-eye-fill"></i></a>
+                            <a href="menu-edit.php?id=<?= $menu['id'] ?>" class="btn btn-link border me-2"><i class="bi bi-pencil-square"></i></a>
+                            <a href="actions/menu-delete.php?id=<?= $menu['id'] ?>" class="btn btn-link border me-2"><i class="bi bi-trash"></i></a>
+                        </p>
+                    </div>
+                <?php endif; ?>
             </div>
         </div>
     <?php endforeach; ?>
